Establish public and private types
-  publicTypes.ts contains types to export
-  privateTypes.ts contains types used internally by the implementation

JIRA: FET-608
