+++ libs/sdk-ui-pivot/src/impl/cell/onCellClick.ts
@@ -1,6 +1,6 @@
 // (C) 2007-2021 GoodData Corporation
 import { TableFacade } from "../../tableFacade";
-import { ICorePivotTableProps } from "../../types";
+import { ICorePivotTableProps } from "../../publicTypes";
 import { CellEvent } from "@ag-grid-community/all-modules";
 import { invariant } from "ts-invariant";
 import { IGridRow } from "../data/resultTypes";
