+++ libs/sdk-ui-pivot/src/impl/structure/colDefTemplates.ts
@@ -1,7 +1,7 @@
 // (C) 2007-2021 GoodData Corporation
 import { ColDef, ValueFormatterParams } from "@ag-grid-community/all-modules";
 import { TableFacade } from "../../tableFacade";
-import { ICorePivotTableProps } from "../../types";
+import { ICorePivotTableProps } from "../../publicTypes";
 import { headerClassFactory } from "./colDefHeaderClass";
 import { AVAILABLE_TOTALS } from "../base/constants";
 import { getMeasureCellFormattedValue, getMeasureCellStyle } from "../cell/cellUtils";
