+++ libs/sdk-ui-pivot/src/impl/structure/headers/AggregationsMenu.tsx
@@ -16,13 +16,13 @@ import noop from "lodash/noop";
 
 import Menu from "../../../menu/Menu";
 import { IOnOpenedChangeParams } from "../../../menu/MenuSharedTypes";
-import { IMenuAggregationClickConfig } from "../../../types";
 import menuHelper from "./aggregationsMenuHelper";
 import AggregationsSubMenu from "./AggregationsSubMenu";
 import { AVAILABLE_TOTALS } from "../../base/constants";
 import { IColumnTotal } from "./aggregationsMenuTypes";
 import { TableDescriptor } from "../tableDescriptor";
 import { isDataColGroup, isDataColLeaf, isDataColRootGroup, isSliceCol } from "../tableDescriptorTypes";
+import { IMenuAggregationClickConfig } from "../../../privateTypes";
 
 /*
  * TODO: same thing is in sdk-ui-ext .. but pivot must not depend on it. we may be in need of some lower-level
