+++ libs/sdk-ui-pivot/src/impl/structure/headers/ColumnGroupHeader.tsx
@@ -2,7 +2,7 @@
 import { IHeaderGroupParams } from "@ag-grid-community/all-modules";
 import React from "react";
 
-import { IMenu } from "../../../types";
+import { IMenu } from "../../../publicTypes";
 
 import HeaderCell, { ALIGN_LEFT, ICommonHeaderParams } from "./HeaderCell";
 import { agColId } from "../tableDescriptorTypes";
