+++ libs/sdk-ui-pivot/src/impl/structure/headers/HeaderCell.tsx
@@ -3,12 +3,13 @@ import React from "react";
 import { IntlShape } from "react-intl";
 import cx from "classnames";
 
-import { IMenu, IMenuAggregationClickConfig } from "../../../types";
+import { IMenu } from "../../../publicTypes";
 import { IOnOpenedChangeParams } from "../../../menu/MenuSharedTypes";
 import AggregationsMenu from "./AggregationsMenu";
 import { HEADER_LABEL_CLASS } from "../../base/constants";
 import { TableDescriptor } from "../tableDescriptor";
 import { IExecutionDefinition, ITotal, SortDirection } from "@gooddata/sdk-model";
+import { IMenuAggregationClickConfig } from "../../../privateTypes";
 
 export type AlignPositions = "left" | "right" | "center";
 export const ALIGN_LEFT = "left";
